MAJ de la gestion d'erreur de l'application, MAJ des methodes CRUD crée precedement et implémentation des nouveaux

+++ src/main/java/will/dev/appStore/controller/UserController.java
@@ -33,7 +33,7 @@ public class UserController {
 
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @PutMapping(path = "{id}", consumes = APPLICATION_JSON_VALUE)
-    public void modifier(@PathVariable int id,@RequestBody User user){
+    public void modifier(@PathVariable Long id,@RequestBody User user){
         this.userService.modifier(id, user);
     }
